Fix int conversion

+++ nncore/engine/hooks/events.py
@@ -91,7 +91,7 @@ class CommandLineWriter(Writer):
 
             if next(engine.model.parameters()).device != torch.device('cpu'):
                 mem = torch.cuda.max_memory_allocated()
-                mem_mb = torch.IntTensor([mem / (1024 * 1024)]).cuda()
+                mem_mb = torch.IntTensor([int(mem / (1024 * 1024))]).cuda()
                 if get_world_size() > 1:
                     dist.reduce(mem_mb, 0, op=dist.ReduceOp.MAX)
                 log += ', memory: {}'.format(mem_mb.item())
